add some tooling to locate and destroy 'zombie' containers.
These containers are not managed by our pool, and can cause issues with blocking ports.  Currently, the 'zombies' are just found and displayed by the server, not destroyed.  It needs to be decided when and why to destroy them.  Finally, a solution should be implemented in the port map to see if we can skip blocked ports.