Preliminary heartbeat on commands for terminals
The terminal window in the jupyter notebook doesn't communicate with the
home server on each command.  Neither do the other various kernels.  The
notebook kernels *do* send checkpoints though.  The checkpoints keep the
container alive appropriately.

The terminal does not do this, thus sessions lasting more than lifetimes
expire, and the learners lose their environment and context.  This
commit adds a pre-command script example.  In order to utilize this, all
codelab/jupyter/etc containers need a single base image.

The heartbeat is defined as a pre-execute script from:

https://github.com/rcaloras/bash-preexec

which runs a wget command to tell the home server the container is
still alive.